+++ src/cli/commands/client/shell.rs
@@ -32,7 +32,7 @@ impl Shell {
             environment.insert("TERM".to_string(), "xterm-256color".to_string());
         }
 
-        // Use provided shell, or determine remote operating system to pick a shell
+        // Use provided shell, use default shell, or determine remote operating system to pick a shell
         let cmd = match cmd.into() {
             Some(cmd) => cmd,
             None => {
@@ -41,7 +41,12 @@ impl Shell {
                     .system_info()
                     .await
                     .context("Failed to detect remote operating system")?;
-                if system_info.family.eq_ignore_ascii_case("windows") {
+
+                // If system reports a default shell, use it, otherwise pick a default based on the
+                // operating system being windows or non-windows
+                if !system_info.shell.is_empty() {
+                    system_info.shell
+                } else if system_info.family.eq_ignore_ascii_case("windows") {
                     "cmd.exe".to_string()
                 } else {
                     "/bin/sh".to_string()