+++ tests/cli/action/system_info.rs
@@ -17,12 +17,20 @@ fn should_output_system_info(mut action_cmd: CtxCommand<Command>) {
                 "Arch: {:?}\n",
                 "Cwd: {:?}\n",
                 "Path Sep: {:?}\n",
+                "Username: {:?}\n",
+                "Shell: {:?}\n",
             ),
             env::consts::FAMILY.to_string(),
             env::consts::OS.to_string(),
             env::consts::ARCH.to_string(),
             env::current_dir().unwrap_or_default(),
             std::path::MAIN_SEPARATOR,
+            whoami::username(),
+            if cfg!(windows) {
+                std::env::var("ComSpec").unwrap_or_else(|_| String::from("cmd.exe"))
+            } else {
+                std::env::var("SHELL").unwrap_or_else(|_| String::from("/bin/sh"))
+            }
         ))
         .stderr("");
 }